+++ src/lib.rs
@@ -52,10 +52,10 @@
 //! use config_manager::{config, ConfigInit};
 //!
 //! const SUFFIX: &str = "_env";
-//!
+//! /// This doc will be included to CLI long_about.
 //! #[derive(Debug)]
 //! #[config(
-//!     clap(version, author),
+//!     clap(version, author, long_about),
 //!     env_prefix = "demo",
 //!     file(
 //!         format = "toml",
@@ -65,7 +65,8 @@
 //!     )
 //! )]
 //! struct MethodConfig {
-//!     #[source(clap(long, short))]
+//!     /// This doc will be included to CLI help.
+//!     #[source(clap(long, short, help))]
 //!     a: i32,
 //!     #[source(
 //!         env(init_from = &format!("b{}", SUFFIX)),